- From Quintijn Hendrickx: I've recently made some changes to osgOcean to improve shoreline rendering. In particular I added a translucency effect at the shores by sampling depth values from the refraction buffer and reconstructing the corresponding world coördinates. The distance between the ocean floor and ocean surface is used to calculate the light extinction level. The constructed world coördinates are also used to add foam along the shorelines.
- Modified the submission slightly so that refraction can still be disabled at run time by the application if necessary.

- Also a small modification to remove ambiguity when multiplying a complex<double> with the result of sqrt (which is overloaded to return float, double or long double as needed).

+++ src/osgOcean/FFTSimulation.cpp
@@ -263,7 +263,7 @@ void FFTSimulation::Implementation::computeBaseAmplitudes()
             K.x() = _PI2*x2*oneOverLen;
 
             RandUtils::gaussianRand(real,imag);
-            _baseAmplitudes[y*(_N+1)+x] = complex(real,imag) * sqrt( 0.5 * phillipsSpectrum(K) );
+            _baseAmplitudes[y*(_N+1)+x] = complex(real,imag) * sqrt( 0.5 * (double)phillipsSpectrum(K) );
         }
     }
 }
